Feature/revizia kodu exception handling

+++ SportApp/Core/src/main/java/com/so/core/services/document/DocumentService.java
@@ -68,14 +68,20 @@ public class DocumentService {
     public void deleteFile(Resource r) throws AppException {
 
         if (r != null) {
-            File file = new File(PATH + "\\" + r.getPath());
-            if (file.delete()) {
-                resourceRepo.delete(r);
-                LOG.info(file.getName() + " subor je vymazany!");
-            } else {
-                LOG.error("Delete operation is failed.");
-                throw new AppException(HttpStatus.INTERNAL_SERVER_ERROR, "nepodarilo sa vymazat resource");
+            File file = new File(PATH + r.getPath());
+            if (file.exists()) {
+                if (file.delete()) {
+                    resourceRepo.delete(r);
+                    LOG.info(file.getName() + " subor je vymazany!");
+                } else {
+                    LOG.error("Delete operation is failed.");
+                    throw new AppException(HttpStatus.INTERNAL_SERVER_ERROR, "nepodarilo sa vymazat resource");
+                }
+            } else{
+                LOG.info("subor {} neexistuje, ale resource bol vymazany",r.getPath());
             }
+        } else {
+            LOG.info("resource je null");
         }
     }
 
@@ -89,6 +95,10 @@ public class DocumentService {
         String filePath = PATH + name;
         try {
             File file = new File(filePath);
+            if(!file.exists()){
+                throw new AppException(HttpStatus.NOT_FOUND,"image:"+name+" nebol najdeny");
+            }
+            
             FileInputStream fis = new FileInputStream(file);
             byte[] fileBytes;
             try (BufferedInputStream inputStream = new BufferedInputStream(fis)) {
@@ -100,10 +110,10 @@ public class DocumentService {
             return r;
         } catch (FileNotFoundException ex) {
             LOG.error("nepodarilo sa nacitat subor={}", name);
-            throw new AppException(HttpStatus.NOT_FOUND, "nepodarilo sa nacitat subor:" + "name " + ex);
+            throw new AppException(HttpStatus.NOT_FOUND, "nepodarilo sa nacitat subor:" + "name " + ex.getMessage());
         } catch (IOException e) {
             LOG.error("nepodarilo sa nacitat subor={}", name);
-            throw new AppException(HttpStatus.NOT_FOUND, "nepodarilo sa nacitat subor:" + "name " + e);
+            throw new AppException(HttpStatus.NOT_FOUND, "nepodarilo sa nacitat subor:" + "name " + e.getMessage());
 
         }
